+++ src/routes/teach/+page.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-    import Header from '@components/app/Header.svelte';
     import { locales } from '@db/Database';
     import MarkupHtmlView from '../../components/concepts/MarkupHTMLView.svelte';
     import { type Class } from '@db/teachers/TeacherDatabase.svelte';
@@ -8,6 +7,7 @@
     import Subheader from '@components/app/Subheader.svelte';
     import { getTeachData } from './+layout.svelte';
     import Spinning from '@components/app/Spinning.svelte';
+    import TeachersOnly from './TeachersOnly.svelte';
 
     let teach = getTeachData();
 
@@ -27,29 +27,30 @@
     >
 {/snippet}
 
-<Header>{$locales.get((l) => l.ui.page.teach.header)}</Header>
-{#if classes === undefined}
-    <Spinning></Spinning>
-{:else if classes === null}
-    <MarkupHtmlView
-        markup={$locales.get((l) => l.ui.page.teach.error.offline)}
-    />
-{:else}
-    {#if classes.length === 0}
+<TeachersOnly>
+    {#if classes === undefined}
+        <Spinning></Spinning>
+    {:else if classes === null}
         <MarkupHtmlView
-            markup={$locales.get((l) => l.ui.page.teach.prompt.none)}
+            markup={$locales.get((l) => l.ui.page.teach.error.offline)}
         />
     {:else}
-        <MarkupHtmlView
-            markup={$locales.get((l) => l.ui.page.teach.prompt.some)}
-        />
+        {#if classes.length === 0}
+            <MarkupHtmlView
+                markup={$locales.get((l) => l.ui.page.teach.prompt.none)}
+            />
+        {:else}
+            <MarkupHtmlView
+                markup={$locales.get((l) => l.ui.page.teach.prompt.some)}
+            />
+        {/if}
+        <Centered>
+            <Link to="/teach/class/new">
+                {$locales.get((l) => l.ui.page.teach.link.new)}
+            </Link>
+        </Centered>
+        {#each classes as group}
+            {@render classDetails(group)}
+        {/each}
     {/if}
-    <Centered>
-        <Link to="/teach/class/new">
-            {$locales.get((l) => l.ui.page.teach.link.new)}
-        </Link>
-    </Centered>
-    {#each classes as group}
-        {@render classDetails(group)}
-    {/each}
-{/if}
+</TeachersOnly>