+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 - Fixed #673, more visible links in paragraphs.
 - Fixed #672, more visible links in feedback with background color.
 - Serialize locale verification.
+- Fixed French translation errors.
+- Fixed Gujarati translation errors.
 
 ## 0.15.0 2024-01-26