+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 - Fixed French translation errors.
 - Fixed Gujarati translation errors.
 - Fixed Hindi translation errors.
+- Fixed Marathi translation errors.
 - Account for full width punctuation in reserved symbols.
 
 ## 0.15.0 2024-01-26