+++ CHANGELOG.md
@@ -26,6 +26,10 @@ Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http:
 - Fixed Tamil translation errors.
 - Account for full width punctuation in reserved symbols.
 
+### Maintenance
+
+- Updated firebase and firebase-functions.
+
 ## 0.15.0 2024-01-26
 
 ### Added